+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Add `__AVR__` to defines when compiling
+- `arduino_ci_remote.rb` CLI switch `--skip-examples-compilation`
 - Add support for `diditalPinToPort()`, `digitalPinToBitMask()`, and `portOutputRegister()`
 
 ### Changed
@@ -15,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Revise math macros to avoid name clashes
 
 ### Deprecated
+- `arduino_ci_remote.rb` CLI switch `--skip-compilation`
 - Deprecated `arduino_ci_remote.rb` in favor of `arduino_ci.rb`
 
 ### Removed